requirements.txt
Flask==1.1.1
numpy==1.18.1
opencv-python==4.2.0.32
python-dotenv==0.10.5
